refactor: Phase 5 slice 5a — SettingsView decomposition (#118)
Split the 3435-line SettingsView.vue (147 top-level fns, 8 tabs) into
components/settings/{SettingsShell,General,Storage,Connectors,Merging,Player,Shortcuts,WatchTogether,Debug}Tab.vue.
Behavior is byte-identical; no IPC, store, or feature changes.

- SettingsShell.vue (206 L) — tab nav + keep-alive <component :is> +
  saved-toast container. Owns tabs/topbar/saved-toast styles.
- 8 tab components hydrate their own settings on first activation, own
  their tab-local broadcast subscriptions (storage:* in StorageTab,
  auto-dl tick in GeneralTab, syncplay status per-test in
  WatchTogetherTab), and dispose on unmount — Phase 4 disposer
  contract preserved (no removeAllListeners/window.api.off* added).
- composables/use-settings-autosave.ts (31 L) — module-level singleton
  for the Saved-toast (savedVisible + showSaved + autoSave). Module
  state keeps the toast a singleton across tabs.
- assets/settings-tabs.css (353 L) — shared utility classes
  (.setting-group/.setting-input/.toggle-row/.merge-all-btn/...)
  imported by each tab via <style scoped src=...> so each component
  gets its own scoped copy; the cascade order and scope id match the
  pre-split single block.
- SettingsView.vue collapses to a 7-line re-export of SettingsShell;
  App.vue route map is unchanged so the 5f folder reorg stays a clean
  rename.
- DESIGN.md updated with the new components/settings/ layout.

Quality gate: typecheck + lint (0 errors) + format:check + 111 tests
+ build + scripts/check-subscription-contract.sh all green.

Part of #84 / Phase 5 (#118).
